Move icon and menu/command factories out of core (Target 3)
Extract the icon-resolution family into a shared IconResolver and move the menu/app-bar/command item factories into the overlay layer, updating all descriptor and overlay call sites. Phase C skipped: flyout factory bodies remain instance-bound to reconciler mounting/update and descriptor bridged lambdas.
