+++ src/Reactor/Core/Reconciler.Update.cs
@@ -1188,120 +1188,5 @@ public sealed partial class Reconciler
     private static string Truncate(string s, int maxLen) =>
         s.Length <= maxLen ? s : s[..maxLen] + "…";
 
-    internal void UpdateMenuFlyoutItems(
-        global::System.Collections.Generic.IList<WinUI.MenuFlyoutItemBase> target,
-        MenuFlyoutItemBase[] oldSource,
-        MenuFlyoutItemBase[] newSource)
-    {
-        int oldCount = oldSource.Length;
-        int newCount = newSource.Length;
-        int shared = Math.Min(oldCount, newCount);
 
-        for (int i = 0; i < shared; i++)
-        {
-            switch (newSource[i])
-            {
-                case MenuFlyoutItemData mfi when target[i] is WinUI.MenuFlyoutItem existing:
-                    existing.Text = mfi.Text;
-                    existing.IsEnabled = mfi.IsEnabled;
-                    existing.Icon = ResolveIcon(mfi.IconElement, mfi.Icon);
-                    if (mfi.AccessKey is not null) existing.AccessKey = mfi.AccessKey;
-                    if (mfi.Description is not null)
-                    {
-                        WinUI.ToolTipService.SetToolTip(existing, mfi.Description);
-                        Microsoft.UI.Xaml.Automation.AutomationProperties.SetHelpText(existing, mfi.Description);
-                    }
-                    existing.Tag = mfi;
-                    break;
-
-                case ToggleMenuFlyoutItemData toggle when target[i] is WinUI.ToggleMenuFlyoutItem toggleItem:
-                    toggleItem.Text = toggle.Text;
-                    toggleItem.IsChecked = toggle.IsChecked;
-                    toggleItem.Icon = ResolveIcon(toggle.IconElement, toggle.Icon);
-                    toggleItem.Tag = toggle;
-                    break;
-
-                case RadioMenuFlyoutItemData radio when target[i] is WinUI.RadioMenuFlyoutItem radioItem:
-                    radioItem.Text = radio.Text;
-                    radioItem.IsChecked = radio.IsChecked;
-                    radioItem.Tag = radio;
-                    break;
-
-                case MenuFlyoutSeparatorData when target[i] is WinUI.MenuFlyoutSeparator:
-                    break; // nothing to update
-
-                case MenuFlyoutSubItemData sub when target[i] is WinUI.MenuFlyoutSubItem subItem:
-                    subItem.Text = sub.Text;
-                    subItem.Icon = ResolveIcon(sub.IconElement, sub.Icon);
-                    // Recursively patch sub-items
-                    var oldSub = oldSource[i] is MenuFlyoutSubItemData oldSubData ? oldSubData.Items : [];
-                    UpdateMenuFlyoutItems(subItem.Items, oldSub, sub.Items);
-                    break;
-
-                default:
-                    // Type mismatch — replace the item
-                    target[i] = CreateMenuFlyoutItem(newSource[i]);
-                    break;
-            }
-        }
-
-        // Remove excess
-        for (int i = oldCount - 1; i >= shared; i--)
-            target.RemoveAt(i);
-
-        // Add new
-        for (int i = shared; i < newCount; i++)
-            target.Add(CreateMenuFlyoutItem(newSource[i]));
-    }
-
-    internal static void UpdateAppBarItems(
-        global::System.Collections.Generic.IList<WinUI.ICommandBarElement> target,
-        AppBarItemBase[]? source)
-    {
-        int newCount = source?.Length ?? 0;
-        int oldCount = target.Count;
-
-        // Update shared range (only update if types match, otherwise replace)
-        int shared = Math.Min(oldCount, newCount);
-        for (int i = 0; i < shared; i++)
-        {
-            if (source is null) continue;
-            switch (source[i])
-            {
-                case AppBarButtonData cmd when target[i] is WinUI.AppBarButton abb:
-                    abb.Label = cmd.Label;
-                    abb.IsEnabled = cmd.IsEnabled;
-                    abb.Icon = ResolveIcon(cmd.IconElement, cmd.Icon);
-                    if (cmd.AccessKey is not null) abb.AccessKey = cmd.AccessKey;
-                    if (cmd.Description is not null)
-                    {
-                        WinUI.ToolTipService.SetToolTip(abb, cmd.Description);
-                        Microsoft.UI.Xaml.Automation.AutomationProperties.SetHelpText(abb, cmd.Description);
-                    }
-                    abb.Tag = cmd;
-                    break;
-                case AppBarToggleButtonData toggle when target[i] is WinUI.AppBarToggleButton atb:
-                    atb.Label = toggle.Label;
-                    atb.IsChecked = toggle.IsChecked;
-                    if (toggle.Icon is not null) atb.Icon = new WinUI.SymbolIcon(ParseSymbol(toggle.Icon));
-                    atb.Tag = toggle;
-                    break;
-                case AppBarSeparatorData when target[i] is WinUI.AppBarSeparator:
-                    break; // nothing to update
-                default:
-                    // Type mismatch — replace
-                    target[i] = CreateAppBarItem(source[i]);
-                    break;
-            }
-        }
-
-        // Remove excess
-        for (int i = oldCount - 1; i >= shared; i--)
-            target.RemoveAt(i);
-
-        // Add new
-        if (source is not null)
-            for (int i = shared; i < newCount; i++)
-                target.Add(CreateAppBarItem(source[i]));
-    }
 }
